Updating zipcode page to be clientside and adding trpc mutation
