+++ src/server/api/routers/zipcode.ts
@@ -4,10 +4,23 @@ import { createTRPCRouter, publicProcedure } from "../trpc";
 
 export const zipcodeRouter = createTRPCRouter({
   postZipCode: publicProcedure
-    .input(z.object({ text: z.string() }))
+    .input(z.object({ userId: z.string(), zipcode: z.string() }))
     .query(({ input }) => {
+      // .mutation(async ({ input, ctx }) => {
+      //   const { name } = ctx.user;
+      //   const post = await prisma.post.create({
+      //     data: {
+      //       ...input,
+      //       name,
+      //       source: 'GITHUB',
+      //     },
+      //   });
+      //   ee.emit('add', post);
+      //   delete currentlyTyping[name];
+      //   ee.emit('isTypingUpdate');
+      //   return post;
       return {
-        zipcode: `User's zipcode: ${input.text}`,
+        zipcode: `User: ${input.userId} zipcode: ${input.zipcode}`,
       };
     }),
